[IMP] Rename js file to 'screens', and return original screens
This 'good practice' improves compatibility with other modules.
