+++ pos_require_product_quantity/readme/CONTRIBUTORS.rst
@@ -1,1 +1,2 @@
 * Robin Keunen <robin.keunen@coopiteasy.be>
+* Iván Todorovich <ivan.todorovich@gmail.com>
