+++ pos_require_product_quantity/views/assets.xml
@@ -2,7 +2,7 @@
 <odoo>
     <template id="assets" name="pos_require_product_quantity_extension_assets" inherit_id="point_of_sale.assets">
       <xpath expr="." position="inside">
-          <script type="text/javascript" src="/pos_require_product_quantity/static/src/js/pos.js"/>
+          <script type="text/javascript" src="/pos_require_product_quantity/static/src/js/screens.js"/>
       </xpath>
     </template>
 </odoo>
